requirements.txt
django>=6.0.0
django-ninja
yt-dlp
gunicorn
whitenoise
